[FIX] sale_project: convert to project currency in profitability
Current behaviour:
All sold services in another currency than the companie's one are not
being converted in the project profitability, but the currency
symbol is correctly set to the one of the project. (related to company)

Expected behaviour:
Sold services in another currency should be converted in the project
profitability.

Steps to reproduce:
- Install Sales, Accounting, Project, and a localization (ex: Sweden)
- Switch to the Swedish company.
- Settings:
  - Activate analytic accounts
  - Make sure the company currency is SEK
  - Update the conversions rates of the currencies
- In products: create a new service (ex: $100)
- Create a new project (make sure it creates an analytic account)
- Sell 1 unit of that service, make sure the pricelist before adding
  the line is "USD Public pricelist" (so the line should be $100),
  add the analytic account to the line. Confirm the SO.
- Add a task in the project, and add the SO as ref to it.
- Go to the project profitability panel, observe that the value of
  the service we just sold is 100 kr (the value of $100, but with
  the symbol for SEK kr).

Reason for the problem:
Lack of currency conversion for services when computing the
profitability items for the project update panel.

Fix:
Similar fix to 68b4ea69c173dd5894bcb42b99e6818f241f9734, but in that
fix we missed the conversion for revenues coming from analytic
accounts (aka services).

Affected versions:
- 16.0
- saas-16.1
- saas-16.2
- master

opw-3274626

closes odoo/odoo#123366

X-original-commit: 9ab317e80e37de129edd488f18e7f2cbcd649e74
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>