[IMP] sale_project: compute the amount invoiced from the down payments for the project profitability
This commit's purpose is to take into account the down payments for the
project profitability of the project update window.
Currently, if an SO has a total of 300$, and a down payment is created for 100$, the project
profitability will not compute the 100$ into the 'amount_invoiced'
column. After this commit, it will. A negative value of -100 is also
added in the 'amount_to_invoice' column in order to keep the total of
that column, and the total of the 'expected' column of the project
update consistent with the total of the SO.

task-3204827

closes odoo/odoo#116293

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>